moves the multicast discovery strategy to com.hazelcast.spi.discovery.multicast to prevent having a user pointing to com.hazelcast.internal.
