+++ hazelcast-client/src/test/java/com/hazelcast/client/spi/impl/discovery/ClientToMemberDiscoveryTest.java
@@ -23,7 +23,7 @@ import com.hazelcast.config.Config;
 import com.hazelcast.config.XmlConfigBuilder;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.Member;
-import com.hazelcast.internal.plugin.multicast.MulticastDiscoveryStrategy;
+import com.hazelcast.spi.discovery.multicast.MulticastDiscoveryStrategy;
 import com.hazelcast.test.AssertTask;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
