+++ hazelcast/src/main/java/com/hazelcast/spi/discovery/multicast/impl/MulticastMemberInfo.java
@@ -14,7 +14,25 @@
  * limitations under the License.
  */
 
-/**
- * Contains the hazelcast multicast plugin classes
- */
-package com.hazelcast.internal.plugin.multicast;
+package com.hazelcast.spi.discovery.multicast.impl;
+
+import java.io.Serializable;
+
+public class MulticastMemberInfo implements Serializable {
+
+    private String host;
+    private int port;
+
+    public MulticastMemberInfo(String host, int port) {
+        this.host = host;
+        this.port = port;
+    }
+
+    public int getPort() {
+        return port;
+    }
+
+    public String getHost() {
+        return host;
+    }
+}
