+++ hazelcast/src/test/resources/hazelcast-multicast-plugin.xml
@@ -30,7 +30,7 @@
             <tcp-ip enabled="false" />
             <aws enabled="false"/>
             <discovery-strategies>
-                <discovery-strategy enabled="true" class="com.hazelcast.internal.plugin.multicast.MulticastDiscoveryStrategy">
+                <discovery-strategy enabled="true" class="com.hazelcast.spi.discovery.multicast.MulticastDiscoveryStrategy">
                 </discovery-strategy>
             </discovery-strategies>
         </join>
